% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{Blast.all}
\alias{Blast.all}
\title{BLAST Single FASTA File}
\usage{
Blast.all(file_name, blast_db, DBname)
}
\arguments{
\item{file_name}{Character. Path to FASTA file}

\item{blast_db}{Character. Path to BLAST database}

\item{DBname}{Character. Database name for output file naming}
}
\value{
NULL. Results are saved to CSV file
}
\description{
Processes a single FASTA file: removes spaces from headers, runs BLAST,
and saves results to CSV.
}
\examples{
\dontrun{
Blast.all(
  file_name = "sample_16S.fasta",
  blast_db = "path/to/16S_ribosomal_RNA",
  DBname = "16S"
)
}
}
