% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mestools-utils.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check Package Dependencies}
\usage{
check_dependencies(packages, auto_install = FALSE)
}
\arguments{
\item{packages}{Character vector of package names}

\item{auto_install}{Logical, if TRUE missing packages will be installed automatically}
}
\value{
Logical indicating if all packages are available
}
\description{
Verifies if all required packages are installed and available.
}
\examples{
check_dependencies(c("ggplot2", "dplyr"))
}
