% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{create_sanger_contig}
\alias{create_sanger_contig}
\title{Generate Sanger Contig from Forward and Reverse Reads}
\usage{
create_sanger_contig(
  path,
  contigName,
  suffixForwardRegExp,
  suffixReverseRegExp,
  file_name_fwd,
  file_name_rev
)
}
\arguments{
\item{path}{Character. Path to directory containing .ab1 files}

\item{contigName}{Character. Name for the contig}

\item{suffixForwardRegExp}{Character. Regular expression for forward read suffix}

\item{suffixReverseRegExp}{Character. Regular expression for reverse read suffix}

\item{file_name_fwd}{Character. Forward read filename}

\item{file_name_rev}{Character. Reverse read filename}
}
\value{
List containing summary statistics and contig object
}
\description{
Creates a consensus sequence from forward and reverse Sanger reads using
the sangeranalyseR package.
}
\examples{
\dontrun{
contig <- create_sanger_contig(
  path = "path/to/files",
  contigName = "Sample01",
  suffixForwardRegExp = "_FWD",
  suffixReverseRegExp = "_REV",
  file_name_fwd = "Sample01_FWD.ab1",
  file_name_rev = "Sample01_REV.ab1"
)
}
}
