% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{edit.fasta}
\alias{edit.fasta}
\title{Edit FASTA File to Remove Spaces}
\usage{
\method{edit}{fasta}(x)
}
\arguments{
\item{x}{Character. Path to FASTA file}
}
\value{
NULL. File is modified in place
}
\description{
Removes spaces from FASTA file headers and replaces them with underscores.
This is necessary for proper BLAST processing.
}
\examples{
\dontrun{
edit.fasta("path/to/sequences.fasta")
}
}
