% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mestools-core-backup.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract Metadata from GEO Object}
\usage{
extract_metadata(meta, sep = ": ")
}
\arguments{
\item{meta}{GEO metadata object}

\item{sep}{Separator for characteristic parsing}
}
\value{
Data frame with cleaned metadata
}
\description{
Helper function to extract metadata from GEO series matrix objects.
}
