% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mestools-utils.R
\name{install_and_load}
\alias{install_and_load}
\title{Install and Load R Packages from Multiple Sources}
\usage{
install_and_load(packages)
}
\arguments{
\item{packages}{A character vector or list of package names.
To install directly from GitHub without searching, use the 'username/reponame' format.}
}
\value{
Loads the requested packages into the current session. Prints status
messages and continues processing even if one package fails.
}
\description{
This function iterates through a vector of package names. If a package is
not installed, it attempts to install it from CRAN, then Bioconductor. If
still not found, it will search GitHub for a repository with that name and
prompt the user to interactively select from the top results for installation.
}
\examples{
# packages <- c("ggplot2", "limma", "tidyverse/dplyr", "hrbrthemes")
# install_and_load(packages)
}
