% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{load_sanger_files}
\alias{load_sanger_files}
\title{Load and Process Sanger Sequencing Files}
\usage{
load_sanger_files(path = path)
}
\arguments{
\item{path}{Character. Path to directory containing .ab1 files}
}
\value{
Character vector of unique group names
}
\description{
Loads ABI files, renames them by removing date/time stamps, and creates
groups of forward and reverse reads.
}
\examples{
\dontrun{
groups <- load_sanger_files("path/to/abi/files")
}
}
