% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mestools-package.R
\docType{package}
\name{mestools}
\alias{mestools-package}
\alias{mestools}
\title{mestools: A Collection of Useful Tools}
\description{
Provides utility functions for data manipulation, package development,
and deployment automation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vanhungtran/mestools}
  \item Report bugs at \url{https://github.com/vanhungtran/mestools/issues}
}

}
\author{
\strong{Maintainer}: Lucas VHH TRAN \email{tranhungydhcm@gmail.com}

}
\keyword{internal}
