% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{print.primer_analysis}
\alias{print.primer_analysis}
\title{Print Primer Analysis}
\usage{
\method{print}{primer_analysis}(x, ...)
}
\arguments{
\item{x}{A primer_analysis object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Primer Analysis
}
