% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{print.primer_order}
\alias{print.primer_order}
\title{Print Primer Order Sheet}
\usage{
\method{print}{primer_order}(x, ...)
}
\arguments{
\item{x}{A primer_order object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Primer Order Sheet
}
