% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{print.primer_recommendation}
\alias{print.primer_recommendation}
\title{Print Primer Recommendation}
\usage{
\method{print}{primer_recommendation}(x, ...)
}
\arguments{
\item{x}{A primer_recommendation object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Primer Recommendation
}
