% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{print_primer_stats}
\alias{print_primer_stats}
\title{Print Primer Statistics}
\usage{
print_primer_stats()
}
\description{
Pretty-print summary of primer database statistics
}
\examples{
\dontrun{
print_primer_stats()
}
}
