% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers-16s.R
\name{setup_blast_env}
\alias{setup_blast_env}
\title{Set BLAST Environment Variables}
\usage{
setup_blast_env(blast_path = "../ncbi-blast-2.13.0+")
}
\arguments{
\item{blast_path}{Character. Path to NCBI BLAST+ installation directory}
}
\value{
NULL. Environment variables are set
}
\description{
Helper function to set up BLAST+ environment variables for the current R session.
Call this before using any BLAST functions.
}
\examples{
\dontrun{
setup_blast_env("path/to/ncbi-blast-2.13.0+")
}
}
